more

+++ src/main/java/ca/jessewhite/imaginationplugin/ai/Block.java
@@ -1,3 +1,3 @@
 package ca.jessewhite.imaginationplugin.ai;
 
-public record Block(String type, int x, int y, int z, boolean fill, Integer endX, Integer endY, Integer endZ) {}
+public record Block(String type, double x, double y, double z, boolean fill, Integer endX, Integer endY, Integer endZ) {}
